Add log method to follow a single container
