+++ package.json
@@ -29,6 +29,7 @@
     "clear": "^0.1.0",
     "clui": "^0.3.6",
     "configstore": "^5.0.1",
+    "docker-cli-js": "^2.7.1",
     "esm": "^3.2.25",
     "execa": "^4.0.0",
     "figlet": "^1.3.0",
